Examples of how to avoid using 'continue'
These are just friendly, suggested refactorings. It shows how to avoid using 'continue' and to limit the number of 'return's, resulting in what I would consider cleaner, more readable, more refactorable and less bug-prone code. 

If it seems like these refactorings result in too many indents, that's a signal that the code could perhaps need further 'extract method' or 'extract function' refactorings. Naming those methods or functions would also be a nice way to comment what is happening. Agreed, function calls are slightly costly, so it shouldn't always be done, especially not when it would result in a lot of function calls inside a 'for' loop with many iterations.

You don't have to agree, it's just a friendly suggestion. :-)